move logger to scipts dir
